Modify framework_objc module to include ringrtc
When building for iOS, the build_ios_libs.py script is used which calls
gn for the framework_objc target. We modify the BUILD.gn configuration
for framework_objc so that it refences the following RingRTC targets as
dependencies:

src/rffi/BUILD.gn
This ends up being a static library linked in with the resulting
WebRTC.framework.