Keep the ringrtc code out of the main WebRTC tree
This is an attempt to keep the ringrtc code out of the main WebRTC
tree, yet still leverage the WebRTC gn/ninja build system.

Step 1 -- Add a generic //ringrtc/BUILD.gn file that depends on an out
of tree BUILD.gn file using an absolute path.

This file includes a <never-checked-in> generated file,
ringrtc-root-dir.gni, that contains the absolute path of the actual
ringrtc source code.  The ringrtc-root-dir.gni file is generated at
the time the work space is created.

Step 2 -- Update the top level //BUILD.gn file to depend on the minimal
//ringrtc/BUILD.gn file.