Modify sdk/objc to support RingRTC requirements
The Objective-C++ interfaces and code are modified to support the
CallConnection classes being defined for RingRTC support. The primary
need is to add an alternative initialization of the RTCPeerConnection
object, which is the main WebRTC interface for iOS.

A new initialization method is provided to allow a custom observer to
be used. In our case, this observer is coming from the Rust RingRTC
code.

The result is that the Rust code implements the observer around the low
level WebRTC object, which is actually instantiated from the iOS side
via Objective-C/C++. The application-level CallConnection object wraps
it all up in a way that is still compatible with most normal WebRTC
operations and objects.