Revert pan to right on MacBook Pro
For native apps, sound on a MacBook Pro is not very good because the
output was pegged to the right speaker if internal speakers are in
use. The reason is that on pretty much all MacBook Pro models, the
microphone is mounted in close proximity to the left speaker.

In the past this fix might have avoided distorted sound or undesirable
AEC performance. However, it seems that on modern models dating back
to 2013 the sound is better overall if not panned to the right,
possibly due to newer AEC capabilities.