Fix loss based BWE to avoid unnecessary paddings.
When the delay-based estimate and a 2% bandwidth-increase candidate had the same kbps but different bps, the algorithm selected the latter.
This resulted in a lower bandwidth selection than the delay-based estimate, triggering unnecessary padding.

With this change, the higher bandwidth candidate will be selected if having the same kbps.

Bug: webrtc:12707,webrtc:427648364
Change-Id: I1ed042689317c768b773d7e6f0819690c38f0e56
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/398080
Commit-Queue: Diep Bui <diepbp@webrtc.org>
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45027}