Introduce FullScreenWindowDetector for Window.Graphics.Capture API
This change defines a FullScreenWindowDetector for window captures while
using WGC on Windows OS. FullScreenWindowDetector detects if an
application exists in a full screen mode and if it does, then it tells
the capturer to capture the full screen version of the application.
Currently this works only for PowerPoint.

This feature is protected behind a finch feature which will be landed in
a follow-up CL in Chromium.

Bug: chromium:409473386
Change-Id: I0f2eeaa5dfb00e76f08ab3a264df193cc643042e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/397640
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Palak Agarwal <agpalak@google.com>
Cr-Commit-Position: refs/heads/main@{#45044}