Deprecate global field trials
On the way delete unused field trials helpers for testing and quering global field trials.

WebRTC is updated to rely on propaged field trials.
When field trials are not propagated it currently fallbacks to the global field trial string, but eventually that should be removed.

Bug: webrtc:42220378
Change-Id: I38d0260f15d85912a1b05faa5a3df98d674634be
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/398701
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45060}