dtls-in-stun: remove restriction on not having certificates
previously, dtls-in-stun was not supported when a certificate was
configured explicitly via generateCertficate. This was done to avoid the
edge case of large RSA certificates (4096/8192 bytes) which caused the
DTLS packet size during the handshake to exceed the available MTU and
led to fragmentation of the certificate flights

This fragmentation is now supported in the code for DTLS-PQC so the
restriction can be removed.

Manually tested with
  https://jsfiddle.net/fippo/19vwyatu/3/
since generating the large RSA certificates in a unit test takes several
seconds (and is generally untested).

Bug: webrtc:367395350
Change-Id: I75a3764c5c1bc7202ad450bda655251803f9344c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/398044
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Philipp Hancke <phancke@meta.com>
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45067}