RTC_CHECK spatial index in correct range: [0-kMaxSpatialLayers)
Currently the spatial index is an int even though it should be non negative. It is also checked in receiver side if spatial_id >= kMaxSpatialLayers, but doesn't protect explicitly from negative values. That in turn expose to out of bounds read/write which could be a security issue.
In this CL I add an RTC_CHECK that those are in-range. Please see discussion in https://webrtc-review.googlesource.com/c/src/+/390060 with my previous attempt to mitigate this, and conclusion to add an RTC_CHECK instead of changing the spatial_id to uint32_t.

Bug: webrtc:411452026
Change-Id: Ie84e8214c99eef38c43a6adafff5df0bf30c4213
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/394180
Commit-Queue: Philipp Hancke <phancke@meta.com>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Sergey Silkin <ssilkin@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45078}