Reland "Introduce NetworkQueue interface"
This reverts commit 833369b1fcc1ce08023cab19c8242eb9268c07c3.
Original cl in the first patchset.

Original cl description:

The purpose of the interface is to allow network simulations to implement their own queueing.

The existing SimulatedNetwork is refactored to use a NetworkQueue interface.
Per default a simple LeayBucket is used that has the same behavior as
SimulatedNetwork today.

Bug: webrtc:42225697
Change-Id: Ic728a5e5f711cf380a38543fac02849da3861ab6
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/399480
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Per Kjellander <perkj@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45119}