Replace webrtc::test::Packet with RtpPacketReceived
To reduce number of classes that represent an RTP packet.

Functionality of simulating payload data that is absent in an RTP dump
is moved into RtpFileSource that reads from such file.
Functionality to parse RED header is moved to the only user -
rtp_analyze

Bug: webrtc:42225366
Change-Id: I8b66cfa0df67a776d9285dcb05da9c0183d6a400
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/400760
Commit-Queue: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45169}