Delete the legacy webrtc_jni namespace
The webrtc_jni namespace is legacy and not referenced anymore from
within WebRTC:
https://source.chromium.org/search?q=webrtc_jni%20filepath:third_party%2Fwebrtc%2F%20-filepath:out%2F%20-filepath:gen%2F&ss=chromium&type=cs

The wiring to the webrtc_jni namespace is from 2018. Probably there is
no more downstream dependencies. Happy if the WebRTC team can flag if
this is indeed the case.

At Meta, we've updated the code to not use these anymore.

Bug: webrtc:434762441
Change-Id: Ib4c35d238fe73a044a510c9a2f901a2e671961fe
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/402640
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Commit-Queue: Boris Tsirkin <btsirkin@meta.com>
Cr-Commit-Position: refs/heads/main@{#45241}