Improve detecting full screen logic for PowerPoint windows
This change adds additional logic where we verify if the visible
rectangle of the full screen window is the same as the display monitor.

This change also introduces checking the window style of the full screen
window to not be an overlapped window. The overlapped window style
includes the minimized or maximized box style, which was the earlier
requirement.

Bug: chromium:409473386
Change-Id: I3fbaf362769ba302f0ee1f6012e6b773ac95b882
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/402742
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Commit-Queue: Palak Agarwal <agpalak@google.com>
Cr-Commit-Position: refs/heads/main@{#45260}