Add transport_name to CandidatePairChangeEvent
The native Candidate objects within a CandidatePairChangeEvent do not
always contain a transport_name (aka mid). That creates issues where the
transport name is required to properly construct IceCandidate objects
for compatibility with other APIs.

This change adds the transport_name to the CandidatePairChangeEvent and
setting it in P2pTransportChannel. Downstream consumers, including the
JNI and Objective-C layers, are updated to use the new field instead of
the one on the candidate.

Bug: webrtc:42233526
Change-Id: Icb411331af61fada3770b4d65002520ae1d18c7d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/403761
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45330}