LNA: Introduce LnaPermissionInterface::ShouldRequestPermission()
Before this CL there was only one method RequestPermission which
performed some synchronous checks e.g. if the address space was less
public, and if necessary asynchronously requested the LNA permission.

This CL introduces ShouldRequestPermission that only performs the
synchronous part of the check. This allows us to centralize the logic
for permission checking into the embedder and more importantly, allows
us to avoid having to post a task for cases where it's not necessary to
request a permission.

Blink change: https://crrev.com/c/6794524

Bug: chromium:421223919
Change-Id: I8edb06f7b604eaae76e5230d404dd19c02d0e24b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/402460
Commit-Queue: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45335}