Provide default implementation for RemoveIceCandidates
The deprecated `RemoveIceCandidates` method was a pure virtual function.
This required all derived classes, including fakes and mocks used in
tests, to provide a boilerplate implementation.

To simplify the codebase, this change provides a default implementation
in the `PeerConnectionInterface`, allowing for the removal of the
now-redundant overrides from various mock and fake test classes,
reducing code clutter.

Bug: webrtc:42233526
Change-Id: I21b9292f1d7cc6b098f7db7519ff927c560c2d85
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/404120
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45367}