Fix declarations that change meaning inside class
According to the C++20 standard [6.4.7], names used inside classes must
refer to the same declaration when re-evaluated in the class scope.
This means that we have to qualify types that use the same names as the
virtual methods.

Bug: webrtc:439743504
Change-Id: I1921b35ec1e3dde02aef098ada427c3cd865c64e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/405100
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Auto-Submit: Marcin Serwin <marcinserwin@google.com>
Reviewed-by: Magnus Flodman <mflodman@webrtc.org>
Commit-Queue: Marcin Serwin <marcinserwin@google.com>
Cr-Commit-Position: refs/heads/main@{#45452}