Make outbound-rtp lifetime spec-compliant (behind flag).
The spec says to delay the creation of the outbound-rtp until the SDP
answer is set. This is equivalent to checking if the currentDirection is
set, since that is one of the side-effects of negotiating an m-section.

Plan B is opted out of this logic because currentDirection is not
applicable to Plan B. Some drive-by comments are added to the
inbound-rtp code paths for symmetry and clarity.

Bug: chromium:406585888
Change-Id: Ib2c7eb99316cd1f82269a756029d19d4c71b0e46
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/406261
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45446}