Add std::optional<RtcpFeedbackType> MediaContentDescription::preferred_rtcp_cc_ack_type()
This method returns the preferred feedback format to use for a certain media type. This is used to decide if CCFB should be enabled or not.
Also, if a feedback type is decided, this is propagated to Audio Send stream in order for the stream to be aware of if it should be part of BW allocation.

Bug: webrtc:440151993, webrtc:383078466
Change-Id: Id3475d5cf547d7f875e0664e46a81e3c6c8c4f44
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/406160
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Per Kjellander <perkj@webrtc.org>
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45443}