Prepare tests for change in RTP stats object lifetimes.
Trying to enable "WebRTC-RTP-Lifetime" by default revealed that the
tests updated in this CL either were not prepared for inbound-rtp being
missing prior to the first packet having been received and crashing or
was using fake objects where packet counters and current direction of
transceivers were not set, resulting in the objects being missing for
the fake scenario being tested.

This CL updates the test or test expectations such that they don't care
if "WebRTC-RTP-Lifetime" is enabled or disabled.

Bug: chromium:406585888
Change-Id: I449a1e1be4cb56cfb68a347ee77cd08d03c443a5
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/406721
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45460}