Implement external scaling/premapping for video
A new API allows webrtc to request the FrameBuffer to prepare for
software readback at the given resolution.
It's done asynchronously because underlying mechanism in chromium is
asynchronous and can take a long time. This way there will be no
artificial bottleneck in the pipeline.

This CL is a no-op itself, as no FrameBuffer implemnetation overrides
the `PrepareMappedBufferAsync()` yet.

Bug: chromium:397485312
Change-Id: I9670c17aa3041ef2107d932e0e6eaefeee851e25
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/406260
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45556}