Propagate content name to ports at construction time, add sanity checks
This sets the Port::content_name (i.e. mid) at construction time and
removes one call to `set_content_name()`. One call to
`set_content_name()` still remains, which is when sessions are pooled
and reused. Because of that, Port::content_name_ can still not be const.

However, this CL improves constness in P2PTransportChannel and adds
several RTC_DCHECKs to ensure the correctness of the content_name field,
including when it belongs to a pooled and not-pooled sessions.

Bug: none
Change-Id: I9730c2ff13efb33d6ad43f9ca152d8fe93929550
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/407281
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45513}