Start removing JsepSessionDescription::Initialize()
This CL starts to decouple direct usage of the JsepSessionDescription
class and encourage the public factory functions that deal with
SessionDescriptionInterface. The Initialize() method is not removed, but
in a few places where we unnecessarily were using JsepSessionDescription
+ Initialize, we now just use the api.

This CL also introduces SdpDeserialize with SessionDescriptionInterface
which will help transition downstream code away from
JsepSessionDescription.

Bug: webrtc:442220720
Change-Id: I80ef4617d9dff20dbc541690932d67505276b444
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/407444
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45535}