Fix Opus stereo decoding of mono FEC to get trivial stereo output.
When a mono Opus encoder sends FEC, and a stereo decoder is used, the decoder can produce non-trivial stereo audio. This change adds logic to ensure that if the last received packet was mono encoded, the decoded output is forced to be trivial stereo when decoding in stereo mode. A new unit test is added to cover this scenario.

Bug: webrtc:376493209
Change-Id: Ibb9790b11f528b736c7a38b74445b3cdf4cf836c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/407803
Reviewed-by: Henrik Lundin <henrik.lundin@webrtc.org>
Commit-Queue: Markus Lindroth <lindroth@google.com>
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45569}