Refactor WebRTC SDP tests and deprecate old APIs
This refactors the SDP unit tests to use `SessionDescriptionInterface`
instead of `JsepSessionDescription` directly.

The `Initialize()` method is being deprecated, as it's an optional
two-step initialization, which can lead to indeterminate state.
Previously the tests used this and called Initialize() multiple times on
the same object, which made the tests harder to reason about. The tests
have been updated to use the `CreateSessionDescription()` factory method
instead and other api/ methods.

The change also:
* Adds a `RTC_DCHECK` to verify that no pre-existing candidates exist before adding new ones.
* Deprecates the `JsepSessionDescription(SdpType type)` constructor.
* Adds a new `add()` method to `IceCandidateCollection` to move candidates from another collection.
* Adds checks to `Clone()` to flag when candidates were not being cloned correctly.
* Significantly (but necessarily) refactors the SDP unit tests for better readability and to remove duplicate code.

Bug: webrtc:442220720
Change-Id: If78b400b4f53195786f7d61f41998e7ea6912163
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/408100
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45589}