Mark deprecated features with [[deprecated]]
This was the result of asking Gemini
"Consider the interfaces in @api. There are comments that indicate that certain methods are deprecated, but the methods don't have [[deprecated]] on them. Make a CL that adds [[deprecated]] to those methods"

followed by a few rounds of reverting and working around errors.

Bug: None
Change-Id: I5d7b1f89fff01277ce96e940aa7340cdfcb728cc
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/407441
Reviewed-by: Tomas Gunnarsson <tommi@webrtc.org>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45623}