Change Java/Jni code to have its own mapping for SessionDescription
Before there was an implicit dependency between the java class and the
strings in SessionDescriptionInterface via use of the
SdpTypeFromString() method. Instead, this special case conversion is
contained to the java/jni implementation and the dependency on
SessionDescriptionInterface strings, removed.

Outside of example code, this is also the last reference in the webrtc
that relied on SdpTypeFromString when creating session description
objects.

Bug: webrtc:442220720
Change-Id: I54f2e3da8df29d168e95c8d45edaa094d24b322e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/409041
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45636}