Move construction of state in WebRtcVoiceEngine from Init to ctor
This allows us to make a few member variables const and remove two. This
also clarifies the split between construction and Init(). A minor change
in this CL is to also rename the legacy send and receive vectors so that
they can't as easily be confused with the other member variables in the
same files that previously had the same name.

Bug: webrtc:42224720
Change-Id: Iab5eb331e6570fb7219653e705fb9791051f1e91
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/410921
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45704}