Add AV1 QP parser.
This AV1 QP parser uses libgav1 to parse the QP. Hence, also including libgav1 in DEPS since it is currently not used in webrtc. This code is only for evaluation and therefore the dependecy will not affect the prod built size.

No-Presubmit: True
Bug: webrtc:358039777
Change-Id: Id4cbae61101ae3343f059c9404b4df4d9f784877
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/409542
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Emil Vardar (xWF) <vardar@google.com>
Cr-Commit-Position: refs/heads/main@{#45713}